Add operation recovery support to async requests
When client asks for asynchronous request processing, the Job is pushed
to the background for asynchronous execution. If the server operation
disrupts in the middle, the operation should be retried later.

This change refactors `jobs.WorkerPool` implementation in such a way
that `jobs.Job` has a `Type` which is expected to have `Executor`
registered in the worker pool. The worker pool then periodically checks
the DB for unfinished and possibly interrupted jobs and re-schedules
them for execution.

This refactoring also changes a bit the way how transactions are
handled. Now the `flow.Transaction` is completely built and stored in
serialized format in `transactions.Transaction` for later sending,
before detaching the job execution.

The `accounts` package use of `transactions` functionality was also
slightly reduced as the account creation is mostly focused around the
flow transaction and benefits from use of `transactions` functionality
for this was quite low. Meanwhile the shared functionality in
`transactions` package was possible to be better encapsulated inside the
`Service`.